Add files via upload
Here we want to implement clustering on "Credit Card" dataset which downloaded from Kaggle . The dataset has different features such as balance amount, purchase behavior, and credit limit. The number of features for this dataset is 18 and it has5951 instances. We want to implement two different clustering models on this dataset. In this way we do some steps which start with data Preprocessing. This step involves handling missing values and try to visualise some features to have better view about dataset. Finding the Number of Clusters is our next step and we try to find optimal number of clusters for our models. In this way we use techniques like the Elbow Method or the Silhouette Score. Then We'll implement two clustering models, K-Means and Hierarchical Clustering and finally we will analyse and interpret the result. We also make a comparison between two methods ad the end of this research.